feat: Support GDAL_NODATA tag and implement NoData filter transform

+++ geotiff-tools/src/main/java/de/hendrikjanssen/geotifftools/rendering/RenderTarget.java
@@ -22,13 +22,13 @@ public class RenderTarget {
 
     public static RenderTarget ofOriginalGeoTiffSize(GeoTiff geoTiff) {
         return new RenderTarget(new Dimension(
-            geoTiff.getMetadata().getWidth(),
-            geoTiff.getMetadata().getHeight()
+            (int) geoTiff.getMetadata().getWidth(),
+            (int) geoTiff.getMetadata().getHeight()
         ));
     }
 
     public RenderTarget scaleUniform(double factor) {
-        
+
         if (Double.isNaN(factor)) {
             throw new IllegalArgumentException("Factor cannot be NaN");
         }
